feat: deletion only for user role and any deletion for admin role

+++ src/houses/mutations/delete-house.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from 'express';
-import ErrorService, { ServerSetupError } from 'services/error-service';
+import UserModel from 'models/user-model';
+import ErrorService, { ForbiddenError, ServerSetupError } from 'services/error-service';
 import HousesModel from '../model';
 import { HouseViewModel } from '../types';
 
@@ -13,9 +14,13 @@ export const deleteHouse: RequestHandler<
 
   try {
     if (id === undefined) throw new ServerSetupError();
-    const house = await HousesModel.getHouse(id);
-    await HousesModel.deleteHouse(id);
+    if (req.authData === undefined) throw new ServerSetupError();
 
+    const user = await UserModel.getUserByEmail(req.authData.email);
+    const house = await HousesModel.getHouse(id);
+    if (user.role !== 'ADMIN' && user.id !== house.owner.id) throw new ForbiddenError();
+
+    await HousesModel.deleteHouse(id);
     res.status(200).json(house);
   } catch (err) {
     const [status, errorResponse] = ErrorService.handleError(err);
